+++ fc_toolbelt/__init__.py
@@ -1,3 +1,3 @@
 # coding: utf-8
 
-__VERSION__ = '0.1.1'
+__VERSION__ = '0.1.2'